+++ lib/src/dialogs/x_dialog.dart
@@ -4,7 +4,11 @@ import "package:flutter/material.dart";
 
 import "../../x_containers.dart";
 
-/// A custom [AlertDialog]
+/// A custom [AlertDialog] matching the style of XContainers.
+///
+/// It features a title and a content widget and two callback buttons.
+/// The buttons are traditionally used for "yes" and "no" behaviors
+/// but can be renamed and will not be displayed if their name is set to [null].
 class XDialog {
   // VARIABLES =================================================================
 
@@ -44,6 +48,12 @@ class XDialog {
   // CONSTRUCTOR ===============================================================
 
   /// Returns an instance of [XDialog] matching the given parameters.
+  ///
+  /// - [title] and [content] are widgets.
+  /// - [cancelText] and [validateText] are the Strings displayed on the buttons.
+  /// - [onCancel] and [onValidate] are the callbacks run when the buttons are pressed.
+  /// - [color] is the background color of the dialog.
+  /// - [backgroundBlur] is the intensity of the gaussian blur applied to stuff that is "behind" the dialog.
   const XDialog({
     required this.title,
     this.content,
@@ -56,6 +66,12 @@ class XDialog {
   });
 
   /// Returns an XDialog displaying only text messages.
+  ///
+  /// - [title] and [content] are Strings.
+  /// - [cancelText] and [validateText] are the Strings displayed on the buttons.
+  /// - [onCancel] and [onValidate] are the callbacks run when the buttons are pressed.
+  /// - [color] is the background color of the dialog.
+  /// - [backgroundBlur] is the intensity of the gaussian blur applied to stuff that is "behind" the dialog.
   factory XDialog.text({
     required String title,
     String? content,
@@ -63,6 +79,7 @@ class XDialog {
     String? validateText = "Okay",
     void Function()? onCancel,
     void Function()? onValidate,
+    Color? color,
     double backgroundBlur = 1,
   }) =>
       XDialog(
@@ -72,6 +89,7 @@ class XDialog {
         validateText: validateText,
         onCancel: onCancel,
         onValidate: onValidate,
+        color: color,
         backgroundBlur: backgroundBlur,
       );
 
@@ -87,6 +105,7 @@ class XDialog {
           sigmaY: backgroundBlur,
         ),
         child: AlertDialog(
+          contentPadding: xTheme.padding,
           backgroundColor: color ?? Theme.of(context).cardColor,
           shape:
               RoundedRectangleBorder(borderRadius: xTheme.dialogBorderRadius),
@@ -99,19 +118,25 @@ class XDialog {
                   Theme.of(context).textTheme.bodyMedium ?? const TextStyle(),
               child: content ?? const SizedBox()),
           actions: <Widget>[
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-                onCancel?.call();
-              },
-              child: Text(cancelText ?? "Cancel"),
+            Visibility(
+              visible: cancelText != null,
+              child: TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                  onCancel?.call();
+                },
+                child: Text(cancelText ?? "Cancel"),
+              ),
             ),
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-                onValidate?.call();
-              },
-              child: Text(validateText ?? "Okay"),
+            Visibility(
+              visible: validateText != null,
+              child: TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                  onValidate?.call();
+                },
+                child: Text(validateText ?? "Okay"),
+              ),
             ),
           ],
         ),